+++ apps/sim/app/api/table/[tableId]/delete-async/route.test.ts
@@ -28,7 +28,7 @@ vi.mock('@sim/utils/id', () => ({
   generateId: vi.fn().mockReturnValue('job-id-xyz'),
   generateShortId: vi.fn().mockReturnValue('short-id'),
 }))
-vi.mock('@/lib/table/service', () => ({
+vi.mock('@/lib/table/jobs/service', () => ({
   markTableJobRunning: mockMarkTableJobRunning,
   releaseJobClaim: mockReleaseJobClaim,
 }))
