+++ apps/sim/app/api/table/[tableId]/export/download/route.ts
@@ -5,7 +5,7 @@ import { parseRequest } from '@/lib/api/server'
 import { checkSessionOrInternalAuth } from '@/lib/auth/hybrid'
 import { generateRequestId } from '@/lib/core/utils/request'
 import { withRouteHandler } from '@/lib/core/utils/with-route-handler'
-import { getTableJob } from '@/lib/table/service'
+import { getTableJob } from '@/lib/table/jobs/service'
 import type { TableExportJobPayload } from '@/lib/table/types'
 import { generatePresignedDownloadUrl } from '@/lib/uploads/core/storage-service'
 import { accessError, checkAccess } from '@/app/api/table/utils'
