+++ apps/sim/lib/table/delete-runner.ts
@@ -5,14 +5,14 @@ import { truncate } from '@sim/utils/string'
 import type { Filter } from '@/lib/table'
 import { TABLE_LIMITS, USER_TABLE_ROWS_SQL_NAME } from '@/lib/table/constants'
 import { appendTableEvent } from '@/lib/table/events'
-import { deletePageByIds, selectRowIdPage } from '@/lib/table/rows/ordering'
 import {
   getJobProgress,
-  getTableById,
   markJobFailed,
   markJobReady,
   updateJobProgress,
-} from '@/lib/table/service'
+} from '@/lib/table/jobs/service'
+import { deletePageByIds, selectRowIdPage } from '@/lib/table/rows/ordering'
+import { getTableById } from '@/lib/table/service'
 import { buildFilterClause } from '@/lib/table/sql'
 
 const logger = createLogger('TableDeleteRunner')
