+++ apps/sim/lib/table/import-runner.ts
@@ -19,16 +19,14 @@ import {
 } from '@/lib/table'
 import { withGeneratedColumnIds } from '@/lib/table/column-keys'
 import { appendTableEvent } from '@/lib/table/events'
+import { markJobFailed, markJobReady, updateJobProgress } from '@/lib/table/jobs/service'
 import { nextImportStartOrderKey, nextImportStartPosition } from '@/lib/table/rows/ordering'
 import {
   addImportColumns,
   bulkInsertImportBatch,
   deleteAllTableRows,
   getTableById,
-  markJobFailed,
-  markJobReady,
   setTableSchemaForImport,
-  updateJobProgress,
 } from '@/lib/table/service'
 import { deleteFile, downloadFileStream, headObject } from '@/lib/uploads/core/storage-service'
 import { normalizeColumn } from '@/app/api/table/utils'
